+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to
 ## Unreleased
 #### Breaking Changes
 #### Added
+- python: add split.py.
 - libvirt: Add pci xml demos.
 - qemu/libvirt: Test cxl persistent,volatile memory.
 - kmod: Add libkmod test directory.